# apis4lawenforcement_swaggerdef.yml
swagger: '2.0'
info:
  version: '1.0'
  title: "APIs 4 SAPOL Demo"
  description: SAPol provides a simple API allowing consumers to view people, license, vehicle and property info.
  #license:
  #  name: MIT
  #  url: https://github.com/apiaryio/polls-api/blob/master/LICENSE
...
basePath: "/"
host: "localhost:3000"
...
schemes:
- http
security:
  - X-MyAPI-Key: []
consumes:
- application/json
produces:
- application/json
paths:
  /tasks:
    x-summary: Tasks 
    post:
      tags: 
      - "Tasks"
      description: >-
        Create a Task representing a scope of work to be actioned. It may or may not be assigned to a Police Officer on createion. The assignment may be done later
      summary: Create a Task
      parameters:
        - name: body
          in: body
          required: true
          schema:
            properties:
              Task:
                $ref: '#/definitions/TaskRequest'
      responses:
        201:
          description: 'Successfully created Task Record'
          schema:
            properties:
              Task:
                $ref: '#/definitions/TaskResponse'
          examples:
            application/json:
              Task:
                taskId: 1
                taskName: 'Serve Summons to appear Ref ABC-1234 to Mr Craig Smith'
                taskDescription: 'Serve Summons to appear Ref ABC-1234 to Mr Craig Smith of 20 Charlesworthnut Street, Walkerville, SA 5000'
                taskLocation: '20 Charlesworthnut Street, Walkerville, SA 5000'
                taskStatus: 'In progress'
                taskStatusUpdateDate : '12-DEC-2018'
                lsa: 'Salisbury'
                officerAssignedToTask: '12345'
                officerAssignedDate: '12-DEC-2018'
                vixenBadgeId: '77777'

    get:
      tags: 
      - "Tasks"
      summary: "Get a List of All Tasks"
      description: "returns a list of tasks "
      operationId: "getTaskInfo"
      produces:
        - "application/json"
      parameters:
      - name: "badgeId"
        in: "query"
        description: "Enter a valid badgeId for an officer"
        required: false
        type: "string"
        x-example: 'A12345'
      - name: "lsa"
        in: "query"
        description: "Enter value of local service area eg Adelaide"
        required: false
        type: "string"
        x-example: 'Adelaide'
      responses:
        200:
          description: Successful Task Lookup Response
          schema:
            properties:
              Tasks:
                type: array
                items:
                  $ref: '#/definitions/TaskResponse'
          examples:
            application/json:
              Tasks:
                - taskId: 1
                  taskName: 'Serve Summons to appear Ref ABC-1234 to Mr Craig Smith'
                  taskDescription: 'Serve Summons to appear Ref ABC-1234 to Mr Craig Smith of 20 Charlesworthnut Street, Walkerville, SA 5000'
                  taskLocation: '20 Charlesworthnut Street, Walkerville, SA 5000'
                  taskStatus: 'In progress'
                  taskStatusUpdateDate : '12-DEC-2018'
                  las: 'Salisbury'
                  officerAssignedToTask: '12345'
                  officerAssignedDate: '12-DEC-2018'
                  vixenBadgeId: '77777'
                - taskId: 2
                  taskName: 'Interview Ms Sharon Smith'
                  taskDescription: 'Interview Ms Sharon Smith of 9 Ann Street, Salisbury, SA 5018 re stolen property - crime ref A3333'
                  taskLocation: '9 Ann Street, Salisbury, SA 5000'
                  taskStatus: 'Assigned'
                  taskStatusUpdateDate : '12-DEC-2018'
                  lsa: 'Salisbury'
                  officerAssignedToTask: '12345'
                  officerAssignedDate: '12-DEC-2018'
                  vixenBadgeId: '77777'
                - taskId: 3
                  taskName: 'Arrest Mike Downer'
                  taskDescription: 'Arrest Mr Mike Downer of 32 Owen Street, Salisbury Downs, SA 5019 re Serious Assualt - crime ref A44444'
                  taskLocation: '32 Owen Street, Salisbury Downs, SA 5109'
                  taskStatus: 'Created'
                  taskStatusUpdateDate : '12-JAN-2019'
                  lsa: 'Salisbury'
                  officerAssignedToTask: '' 
                  officerAssignedDate: ''
                  vixenBadgeId: '77777'
                - taskId: 4
                  taskName: 'Arrest Peter Pan'
                  taskDescription: 'Arrest Mr Peter Pan 25 Grenfell Street, Adelaide, SA 5000 re Serious Assualt - crime ref A44444'
                  taskLocation: '25 Grenfell Street, Adelaide, SA 5000'
                  taskStatus: 'Created'
                  taskStatusUpdateDate : '13-JAN-2019'
                  lsa: 'Adelaide'
                  officerAssignedToTask: '' 
                  officerAssignedDate: ''
                  vixenBadgeId: '88888'                
  /tasks/{taskId}:
    put:
      tags: 
      - "Tasks"
      summary: "Update Task Details"
      description: "Updates a single Task"
      operationId: "updateTaskByTaskId"
      produces:
      - "application/json"
      parameters:
        - name: taskId
          in: path
          required: true
          type: number
          x-example: 27
        - name: body
          in: body
          required: true
          schema:
            properties:
              Task:
                $ref: '#/definitions/TaskRequest'
      responses:
        201:
          description: 'Successfully updated Task Information'
          schema:
            $ref: '#/definitions/TaskResponse'
          examples:
            application/json:
                taskId: 99
                taskName: 'Record Victim of Crime Statement for Court Case - Ms Sally Nyugen - Ref CA1245'
                taskDescription: 'Interview Ms Sally Nguyen of 30 Waterloo Corner Road, Salisbury, SA 5018 re stolen property - crime ref A3333'
                officerAssignedToTask: '12345'
                officerAssignedDate: '12-DEC-2018'
                taskLocation: '9 Ann Street, Salisbury, SA 5000'
                taskStatus: 'Completed'
                taskLastUpdatedDate: '12-DEC-2018'
                lsa: 'Salisbury'
                vixenBadgeId: '77777'
...
securityDefinitions:
  X-MyAPI-Key:
    type: "apiKey"
    name: "x-api-key"
    in: "header"
definitions:
  TaskRequest:
    title: "Task Request"
    type: "object"
    required:
    - "taskName"
    - "taskDescription"
    properties:
      taskName:
        type: string
        example: 'Serve Summons'
      taskDescription:
        type: string
        example: 'Detailed description of the task'
      officerAssignedToTask:
        type: string
        example: 'A12345'
      officerAssignedDate:
        type: string
        example: '20-JAN-2018'
      taskLocation:
        type: string
        example: 'Goodwood, SA'
      taskStatus:
        type: string
        example: 'Open'
      taskLastUpdatedDate:
        type: string
        example: '20-FEB-2018'
      lsa:
        type: string
        example: 'Adelaide'
      vixenBadgeId:
        type: string
        example: 'A77777'
  TaskResponse:
    title: "Task Response"
    type: object
    properties:
      taskId: 
        type: number
        example: 1
      taskName:
        type: string
        example: 'Serve Summons'
      taskDescription:
        type: string
        default: 'Detailed description of the task'
      officerAssignedToTask:
        type: string
        example: 'A12345'
      officerAssignedDate:
        type: string
        example: '20-JAN-2018'
      taskLocation:
        type: string
        example: 'Goodwood, SA'
      taskStatus:
        type: string
        example: 'Open'
      taskLastUpdatedDate:
        type: string
        example: '20-FEB-2018'
      lsa:
        type: string
        example: 'Adelaide'
      vixenBadgeId:
        type: string
        example: 'A77777'
...